requirements.txt
APScheduler==3.9.1
ddddocr==1.4.7
logzero==1.7.0
Pillow==9.2.0
pyOpenSSL==22.1.
requests==2.28.1
six==1.16.0
urllib3==1.26.12